% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doseResponse.R
\name{doseResp}
\alias{doseResp}
\title{Methacholine Dose Response Curve}
\usage{
doseResp(data, respVar)
}
\arguments{
\item{data}{Takes a dataframe with columns c(Mouse, MCh, variable, value) as
input}

\item{respVar}{A string containing the respiratory variable you want eg. 'Rn'}
}
\value{
A linear regression model.
}
\description{
Methacholine Dose Response Curve
}
\examples{
if (requireNamespace("magrittr", quietly = TRUE)) {
summary(doseResp(df))
}
}
