% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmmcv.R
\name{mmmcv}
\alias{mmmcv}
\title{Max Min Mean CV}
\usage{
mmmcv(data, measurements)
}
\arguments{
\item{data}{Takes a dataframe with columns c(Mouse, MCh, variable, value) as
input}

\item{measurements}{A vector containing one or more of the following:
c('max', 'min', 'mean', 'cv')}
}
\value{
A dataframe in long form containing the selected measurements for
  each respiratory variable, mouse, and methacholine dose.
}
\description{
Calculates the maximum, minimum, mean, or percent coefficient of variance of
a respiratory variable value
}
