% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk2tr.R
\name{pk2tr}
\alias{pk2tr}
\title{Peak To Trough}
\usage{
pk2tr(data)
}
\arguments{
\item{data}{Takes a dataframe with columns c(Mouse, MCh, variable, value) as
input}
}
\value{
A dataframe containing the percent decrease from the peak of a
  respiratory variable to the next minimum.
}
\description{
Finds the difference between the maximum of a respiratory value and the
minimum following the peak.
}
