% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provoConc.R
\name{provoConc}
\alias{provoConc}
\title{Provocative Concentration}
\usage{
provoConc(data, respVar, threshold)
}
\arguments{
\item{data}{Takes a dataframe with columns c(Mouse, MCh, variable, value) as
input}

\item{respVar}{A string containing the respiratory variable you want eg. 'Rn'}

\item{threshold}{An integer value of the desired threshold. Eg. a PD20 would
have a threshold of 1.2}
}
\value{
An integer value of the provocative concentration
}
\description{
Calculates the methacholine dose required to change a respiratory variable
past a given threshold.
}
\section{Citation}{
 \code{Sugget, J. A., Nagel, M. W., & Mitchell, J. P.
  (n.d.). The Methacholine Challenge Test for Reversible Airways Disease
  Assessment: A Practical Guide on How to Interpret New 2017 ERS Guidelines.
  Retrieved from https://www.trudellmed.com/node/227}
}

\examples{
provoConc(df, 'Rn', 1.2)
provoConc(df, 'Rn', 2.0)
}
