% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tHalf.R
\name{tHalf}
\alias{tHalf}
\title{T-Half}
\usage{
tHalf(data, timeInterval)
}
\arguments{
\item{data}{Takes a dataframe with columns c(Mouse, MCh, variable, value) as
input}

\item{timeInterval}{The time interval between observations.}
}
\value{
A dataframe containing the T-Half for each respiratory variable of
  each mouse of each methacholine insult.
}
\description{
Calculates the time it takes a respiratory variable to go from a peak after a
methacholine insult to a minimum. Uses a self-start nonlinear regression
model. It will spit out a ton of stuff during use; don't worry about it.
}
